requirements.txt
# Document Processing
python-docx==0.8.11
PyPDF2==3.0.1
langchain==0.1.0
unstructured==0.10.30

# Vector Database
chromadb==0.4.18

# Embeddings & LLM
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.19.0

# Web Interface
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.26.2